Services

+++ src/main/java/com/adventure/adventurexp/Service/ActivityService.java
@@ -24,7 +24,7 @@ public class ActivityService {
         return activityList;
     }
 
-    public Activity getActivityById(Long id) {
+    public Activity getActivityById(Integer id) {
         return activityRepo.findById(id).get();
     }
 
@@ -32,8 +32,8 @@ public class ActivityService {
         return activityRepo.save(activity);
     }
 
-    public void deleteActivity(Long id) {
-        Long count = activityRepo.countById(id);
+    public void deleteActivity(Integer id) {
+        Integer count = activityRepo.countById(id);
         activityRepo.deleteById(id);
     }
 }
